Add fetcher import declaration
